Added parent class LineGroup and moved some methods from Shape
